Refactor Components and update Theme
