+++ docker-compose.yml
@@ -8,6 +8,7 @@ services:
       - 3000:3000/tcp
     environment:
       - GF_AUTH_ANONYMOUS_ORG_ROLE=Admin
+      - GF_EXPLORE_ENABLED=true
       - GF_AUTH_ANONYMOUS_ENABLED=true
       - GF_AUTH_BASIC_ENABLED=false
       - GF_DEFAULT_APP_MODE=development
